refactor(query): split share grant interpreters
