+++ src/query/service/src/interpreters/access/mod.rs
@@ -20,3 +20,5 @@ pub use accessor::AccessChecker;
 pub use accessor::Accessor;
 pub use management_mode_access::ManagementModeAccess;
 pub use privilege_access::PrivilegeAccess;
+pub(crate) use privilege_access::validate_share_management_for_connection;
+pub(crate) use privilege_access::validate_share_object_by_id;
