+++ src/query/service/src/interpreters/interpreter_share.rs
@@ -15,7 +15,6 @@
 use std::collections::BTreeSet;
 use std::sync::Arc;
 
-use databend_common_catalog::database::Database;
 use databend_common_config::GlobalConfig;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
@@ -25,6 +24,7 @@ use databend_common_expression::types::StringType;
 use databend_common_management::RoleApi;
 use databend_common_meta_app::principal::OwnershipObject;
 use databend_common_meta_app::schema::CreateDatabaseReq;
+use databend_common_meta_app::schema::CreateOption;
 use databend_common_meta_app::schema::DatabaseMeta;
 use databend_common_meta_app::schema::database_name_ident::DatabaseNameIdent;
 use databend_common_sql::plans::AlterSharePlan;
@@ -33,28 +33,22 @@ use databend_common_sql::plans::CreateDatabaseFromSharePlan;
 use databend_common_sql::plans::CreateSharePlan;
 use databend_common_sql::plans::DescSharePlan;
 use databend_common_sql::plans::DropSharePlan;
-use databend_common_sql::plans::GrantSharePlan;
-use databend_common_sql::plans::RevokeSharePlan;
-use databend_common_sql::plans::ShareGrantObject;
-use databend_common_sql::plans::ShareGrantObjectPrivilege;
 use databend_common_sql::plans::ShowSharesPlan;
 use databend_common_users::RoleCacheManager;
 use databend_common_users::UserApiProvider;
 use databend_meta_client::types::MatchSeq;
 
 use crate::interpreters::Interpreter;
-use crate::interpreters::access::PrivilegeAccess;
+use crate::interpreters::access::validate_share_management_for_connection;
 use crate::pipelines::PipelineBuildResult;
 use crate::sessions::QueryContext;
 use crate::sessions::TableContextAuthorization;
 use crate::sessions::TableContextTableAccess;
-use crate::share::ProviderObjectIds;
 use crate::share::SHARE_ENGINE;
 use crate::share::SetShareRequest;
-use crate::share::ShareGrantDatabase;
-use crate::share::ShareGrantTable;
 use crate::share::ShareMgr;
 use crate::share::ensure_provider_table_can_be_shared;
+use crate::share::resolve_share_storage_params;
 
 fn share_mgr() -> ShareMgr {
     ShareMgr::create(UserApiProvider::instance().get_meta_store_client())
@@ -85,10 +79,10 @@ impl Interpreter for CreateShareInterpreter {
         let service = share_mgr();
         if service.exists(&self.plan.tenant, &self.plan.name).await? {
             match self.plan.create_option {
-                databend_common_meta_app::schema::CreateOption::CreateIfNotExists => {
+                CreateOption::CreateIfNotExists => {
                     return Ok(PipelineBuildResult::create());
                 }
-                databend_common_meta_app::schema::CreateOption::Create => {
+                CreateOption::Create => {
                     service
                         .create_share(
                             &self.plan.tenant,
@@ -100,15 +94,11 @@ impl Interpreter for CreateShareInterpreter {
                         .await?;
                     return Ok(PipelineBuildResult::create());
                 }
-                databend_common_meta_app::schema::CreateOption::CreateOrReplace => {}
+                CreateOption::CreateOrReplace => {}
             }
         }
         if let Some(connection) = &self.plan.connection {
-            PrivilegeAccess::validate_share_management_for_connection(
-                self.ctx.clone(),
-                Some(connection),
-            )
-            .await?;
+            validate_share_management_for_connection(self.ctx.clone(), Some(connection)).await?;
             UserApiProvider::instance()
                 .get_connection(&self.plan.tenant, connection)
                 .await?;
@@ -173,11 +163,7 @@ impl AlterShareInterpreter {
         service: &ShareMgr,
         connection: &str,
     ) -> Result<BTreeSet<u64>> {
-        PrivilegeAccess::validate_share_management_for_connection(
-            self.ctx.clone(),
-            Some(connection),
-        )
-        .await?;
+        validate_share_management_for_connection(self.ctx.clone(), Some(connection)).await?;
         UserApiProvider::instance()
             .get_connection(&self.plan.tenant, connection)
             .await?;
@@ -201,12 +187,7 @@ impl AlterShareInterpreter {
                 .storage_params
                 .clone()
                 .unwrap_or_else(|| GlobalConfig::instance().storage.params.clone());
-            crate::share::resolve_share_storage_params(
-                &self.plan.tenant,
-                connection,
-                storage_params,
-            )
-            .await?;
+            resolve_share_storage_params(&self.plan.tenant, connection, storage_params).await?;
         }
         Ok(table_ids)
     }
@@ -215,11 +196,7 @@ impl AlterShareInterpreter {
         let connection = service
             .get_connection_name(&self.plan.tenant, &self.plan.name)
             .await?;
-        PrivilegeAccess::validate_share_management_for_connection(
-            self.ctx.clone(),
-            Some(&connection),
-        )
-        .await?;
+        validate_share_management_for_connection(self.ctx.clone(), Some(&connection)).await?;
         Ok(connection)
     }
 }
@@ -297,234 +274,6 @@ impl Interpreter for AlterShareInterpreter {
             }
         }
         Ok(PipelineBuildResult::create())
-    }
-}
-
-pub struct GrantShareInterpreter {
-    ctx: Arc<QueryContext>,
-    plan: GrantSharePlan,
-}
-
-impl GrantShareInterpreter {
-    pub fn try_create(ctx: Arc<QueryContext>, plan: GrantSharePlan) -> Result<Self> {
-        Ok(Self { ctx, plan })
-    }
-}
-
-#[async_trait::async_trait]
-impl Interpreter for GrantShareInterpreter {
-    fn name(&self) -> &str {
-        "GrantShareInterpreter"
-    }
-
-    fn is_ddl(&self) -> bool {
-        true
-    }
-
-    async fn execute2(&self) -> Result<PipelineBuildResult> {
-        match (&self.plan.privilege, &self.plan.object) {
-            (ShareGrantObjectPrivilege::Usage, ShareGrantObject::Database { database }) => {
-                let db = provider_database(&self.ctx, &self.plan.tenant, database).await?;
-                ensure_database_can_be_shared(db.as_ref())?;
-                let grant = ShareGrantDatabase {
-                    database: database.clone(),
-                    database_id: db.get_db_info().database_id.db_id,
-                    database_meta_seq: db.get_db_info().meta.seq,
-                };
-                PrivilegeAccess::validate_share_database_grant(self.ctx.clone(), &grant).await?;
-                share_mgr()
-                    .grant_database(&self.plan.tenant, &self.plan.share, grant)
-                    .await?;
-            }
-            (ShareGrantObjectPrivilege::Select, ShareGrantObject::Table { database, table }) => {
-                let database = self.resolve_table_database(database.as_deref())?;
-                let db = provider_database(&self.ctx, &self.plan.tenant, &database).await?;
-                ensure_database_can_be_shared(db.as_ref())?;
-                let table_ref = db.get_table(table).await?;
-                ensure_provider_table_can_be_shared(&table_ref.get_table_info().meta)?;
-                let storage_params = table_ref
-                    .get_table_info()
-                    .meta
-                    .storage_params
-                    .clone()
-                    .unwrap_or_else(|| GlobalConfig::instance().storage.params.clone());
-                let grant = ShareGrantTable {
-                    database,
-                    database_id: db.get_db_info().database_id.db_id,
-                    database_meta_seq: db.get_db_info().meta.seq,
-                    table: table.clone(),
-                    table_id: table_ref.get_id(),
-                    table_meta_seq: table_ref.get_table_info().ident.seq,
-                };
-                PrivilegeAccess::validate_share_table_grant(self.ctx.clone(), &grant).await?;
-                let manager = share_mgr();
-                let connection = manager
-                    .get_connection_name(&self.plan.tenant, &self.plan.share)
-                    .await?;
-                PrivilegeAccess::validate_share_management_for_connection(
-                    self.ctx.clone(),
-                    Some(&connection),
-                )
-                .await?;
-                crate::share::resolve_share_storage_params(
-                    &self.plan.tenant,
-                    &connection,
-                    storage_params,
-                )
-                .await?;
-                manager
-                    .grant_table(&self.plan.tenant, &self.plan.share, grant, connection)
-                    .await?;
-            }
-            _ => {
-                return Err(ErrorCode::BadArguments(
-                    "Only USAGE ON DATABASE and SELECT ON TABLE can be granted to a share",
-                ));
-            }
-        }
-
-        Ok(PipelineBuildResult::create())
-    }
-}
-
-impl GrantShareInterpreter {
-    fn resolve_table_database(&self, database: Option<&str>) -> Result<String> {
-        resolve_table_database(&self.ctx, database)
-    }
-}
-
-pub struct RevokeShareInterpreter {
-    ctx: Arc<QueryContext>,
-    plan: RevokeSharePlan,
-}
-
-impl RevokeShareInterpreter {
-    pub fn try_create(ctx: Arc<QueryContext>, plan: RevokeSharePlan) -> Result<Self> {
-        Ok(Self { ctx, plan })
-    }
-}
-
-#[async_trait::async_trait]
-impl Interpreter for RevokeShareInterpreter {
-    fn name(&self) -> &str {
-        "RevokeShareInterpreter"
-    }
-
-    fn is_ddl(&self) -> bool {
-        true
-    }
-
-    async fn execute2(&self) -> Result<PipelineBuildResult> {
-        let manager = share_mgr();
-        match (&self.plan.privilege, &self.plan.object) {
-            (ShareGrantObjectPrivilege::Usage, ShareGrantObject::Database { database }) => {
-                let current_database_id =
-                    provider_database_id_if_exists(&self.ctx, &self.plan.tenant, database).await?;
-                if let Some(target) = manager
-                    .prepare_revoke_database(
-                        &self.plan.tenant,
-                        &self.plan.share,
-                        current_database_id,
-                    )
-                    .await?
-                {
-                    PrivilegeAccess::validate_share_revoke_target(self.ctx.clone(), &target)
-                        .await?;
-                    manager
-                        .revoke_share_object(&self.plan.tenant, &self.plan.share, target)
-                        .await?;
-                }
-            }
-            (ShareGrantObjectPrivilege::Select, ShareGrantObject::Table { database, table }) => {
-                let database = resolve_table_database(&self.ctx, database.as_deref())?;
-                let current_object_ids =
-                    provider_table_id_if_exists(&self.ctx, &self.plan.tenant, &database, table)
-                        .await?;
-                if let Some(target) = manager
-                    .prepare_revoke_table(&self.plan.tenant, &self.plan.share, current_object_ids)
-                    .await?
-                {
-                    PrivilegeAccess::validate_share_revoke_target(self.ctx.clone(), &target)
-                        .await?;
-                    manager
-                        .revoke_share_object(&self.plan.tenant, &self.plan.share, target)
-                        .await?;
-                }
-            }
-            _ => {
-                return Err(ErrorCode::BadArguments(
-                    "Only USAGE ON DATABASE and SELECT ON TABLE can be revoked from a share",
-                ));
-            }
-        }
-
-        Ok(PipelineBuildResult::create())
-    }
-}
-
-async fn provider_database(
-    ctx: &QueryContext,
-    tenant: &databend_common_meta_app::tenant::Tenant,
-    database: &str,
-) -> Result<Arc<dyn Database>> {
-    let catalog = ctx.get_default_catalog()?;
-    catalog.get_database(tenant, database).await
-}
-
-async fn provider_database_id_if_exists(
-    ctx: &QueryContext,
-    tenant: &databend_common_meta_app::tenant::Tenant,
-    database: &str,
-) -> Result<Option<u64>> {
-    match provider_database(ctx, tenant, database).await {
-        Ok(database) => Ok(Some(database.get_db_info().database_id.db_id)),
-        Err(err) if err.code() == ErrorCode::UNKNOWN_DATABASE => Ok(None),
-        Err(err) => Err(err),
-    }
-}
-
-async fn provider_table_id_if_exists(
-    ctx: &QueryContext,
-    tenant: &databend_common_meta_app::tenant::Tenant,
-    database: &str,
-    table: &str,
-) -> Result<Option<ProviderObjectIds>> {
-    let database = match provider_database(ctx, tenant, database).await {
-        Ok(database) => database,
-        Err(err) if err.code() == ErrorCode::UNKNOWN_DATABASE => return Ok(None),
-        Err(err) => return Err(err),
-    };
-    let database_id = database.get_db_info().database_id.db_id;
-    match database.get_table(table).await {
-        Ok(table) => Ok(Some(ProviderObjectIds {
-            database_id,
-            table_id: table.get_id(),
-        })),
-        Err(err) if err.code() == ErrorCode::UNKNOWN_TABLE => Ok(None),
-        Err(err) => Err(err),
-    }
-}
-
-fn ensure_database_can_be_shared(db: &dyn Database) -> Result<()> {
-    if db.engine().eq_ignore_ascii_case(SHARE_ENGINE) {
-        return Err(ErrorCode::InvalidOperation(
-            "Cannot grant a shared database to a share",
-        ));
-    }
-    Ok(())
-}
-
-fn resolve_table_database(ctx: &QueryContext, database: Option<&str>) -> Result<String> {
-    match database {
-        Some(database) => Ok(database.to_string()),
-        None => {
-            let database = ctx.get_current_database();
-            if database.is_empty() {
-                Err(ErrorCode::UnknownDatabase("No database selected"))
-            } else {
-                Ok(database)
-            }
-        }
     }
 }
