+++ src/query/service/src/interpreters/mod.rs
@@ -122,6 +122,7 @@ mod interpreter_set;
 mod interpreter_set_priority;
 mod interpreter_set_workload_group_quotas;
 mod interpreter_share;
+mod interpreter_share_grant;
 mod interpreter_show_online_nodes;
 mod interpreter_show_warehouses;
 mod interpreter_show_workload_groups;
@@ -303,9 +304,9 @@ pub use interpreter_share::CreateDatabaseFromShareInterpreter;
 pub use interpreter_share::CreateShareInterpreter;
 pub use interpreter_share::DescShareInterpreter;
 pub use interpreter_share::DropShareInterpreter;
-pub use interpreter_share::GrantShareInterpreter;
-pub use interpreter_share::RevokeShareInterpreter;
 pub use interpreter_share::ShowSharesInterpreter;
+pub use interpreter_share_grant::GrantShareInterpreter;
+pub use interpreter_share_grant::RevokeShareInterpreter;
 pub use interpreter_show_public_keys::ShowPublicKeysInterpreter;
 pub use interpreter_stream_create::CreateStreamInterpreter;
 pub use interpreter_stream_drop::DropStreamInterpreter;
